treewide: Begin GStreamer-based encoder work
Since the MediaCodecSource-based approach seems a little fragile
on newer ports at the moment, and to enable the same kind of
functionality on mainline devices down the road, make encoding
frames work using GStreamer.

Right now it is expected to use gst-droid, but this could be
configured as necessary on mainline devices as well.

- Exchange the default encoder with a GStreamer-based one
- Disable tests for now (will be tackled before proposal)